% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_clin.R
\name{input_clin}
\alias{input_clin}
\title{Preparing input files}
\usage{
input_clin(file_path)
}
\arguments{
\item{file_path}{character string indicating a tab delimited clinical file path}
}
\value{
z data frame of selected clinical data
}
\description{
Preparing input files
}
\examples{
clin_df <- input_clin("/DATA1/lncRNACNVIntegrateR/extdata/clinical_data_COAD.txt.gz")
}
