% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cnv.R
\name{input_cnv}
\alias{input_cnv}
\title{Preparing input files}
\usage{
input_cnv(file_path)
}
\arguments{
\item{file_path}{character string indicating a tab delimited CNV file path}
}
\value{
y data frame of absolute CNV values
}
\description{
Preparing input files
}
\examples{
cnv_df <- input_cnv("/DATA1/lncRNACNVIntegrateR/extdata/CNV_data_COAD.txt.gz")
}
