% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_expr.R
\name{input_expr}
\alias{input_expr}
\title{Preparing input files}
\usage{
input_expr(file_path)
}
\arguments{
\item{file_path}{character string indicating a tab delimited expression file path}
}
\value{
x data frame of expression values

 @import utils
}
\description{
Preparing input files
}
\examples{
expr_df <- input_expr("/DATA1/lncRNACNVIntegrateR/extdata/gene_expression_COAD.txt.gz")
}
