% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_lncRNA_positions.R
\name{input_lncRNA_positions}
\alias{input_lncRNA_positions}
\title{Reading the lncRNA positions in the genome}
\usage{
input_lncRNA_positions()
}
\value{
l data frame of lncRNA positions
}
\description{
Reading the lncRNA positions in the genome
}
\examples{
lncRNA_positions <- input_lncRNA_positions()
}
