% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_and_preparation.R
\name{preprocessing_and_preparation}
\alias{preprocessing_and_preparation}
\title{Preprocessing of input files}
\usage{
preprocessing_and_preparation(expr_df, cnv_df, clin_df)
}
\arguments{
\item{expr_df}{data frame of expression values}

\item{cnv_df}{data frame of absolute CNV values}

\item{clin_df}{data frame of clinical values}
}
\value{
preprocessing_and_preparation_result
}
\description{
Preprocessing of input files
}
\examples{
preprocessing_and_preparation_result <- preprocessing_and_preparation(expr_df, cnv_df, clin_df)
}
