% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_of_lncRNA_and_CNVs_association.R
\name{visualization_of_lncRNA_and_CNVs_association}
\alias{visualization_of_lncRNA_and_CNVs_association}
\title{Visulalization of correlation between lncRNA expression and CNVs profiles}
\usage{
visualization_of_lncRNA_and_CNVs_association(
  result_surv = NULL,
  lncRNA_expression_data_surv_related = NULL,
  lncRNAs_surv_related_CNV_details = NULL
)
}
\arguments{
\item{result_surv}{list returned by extract_survival_related_significant_lncRNA}

\item{lncRNA_expression_data_surv_related}{data frame of expression values}

\item{lncRNAs_surv_related_CNV_details}{data frame of lncRNAs corresponding CNV details}
}
\value{
result_visualization
}
\description{
Visulalization of correlation between lncRNA expression and CNVs profiles
}
\examples{
# Example 1: using the results from previous function:
# Assuming result_surv is obtained from the extract_survival_related_significant_lncRNA function
result_visualization <- visualization_of_lncRNA_and_CNVs_association(result_surv = result_surv)

# Example 2: providing the data frames directly:
# Assuming lncRNA_expression_data_surv_related_df and lncRNAs_surv_related_CNV_details_df are available
result_visualization <- visualization_of_lncRNA_and_CNVs_association(lncRNA_expression_data_surv_related = lncRNA_expression_data_surv_related_df, lncRNAs_surv_related_CNV_details = lncRNAs_surv_related_CNV_details_df)
}
