% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation_lncRNAs_and_CNVs.R
\name{Correlation_lncRNAs_and_CNVs}
\alias{Correlation_lncRNAs_and_CNVs}
\title{Finding correlation between lncRNA expression and CNVs}
\usage{
Correlation_lncRNAs_and_CNVs(
  preprocessing_and_preparation_result = NULL,
  lncRNA_expressions = NULL,
  lncRNA_CNV_profiles = NULL
)
}
\arguments{
\item{preprocessing_and_preparation_result}{list returned by preprocessing_and_preparation function}

\item{lncRNA_expressions}{data frame of expression values}

\item{lncRNA_CNV_profiles}{data frame of absolute CNV values}
}
\value{
Correlation_lncRNAs_and_CNVs_results
}
\description{
Finding correlation between lncRNA expression and CNVs
}
\examples{
# Example 1: Using the result from the preprocessing_and_preparation function
Correlation_lncRNAs_and_CNVs_results <- Correlation_lncRNAs_and_CNVs(preprocessing_and_preparation_result)
# Example 2: Providing lncRNA expression and CNV profiles directly
Correlation_lncRNAs_and_CNVs_results <- Correlation_lncRNAs_and_CNVs(lncRNA_expressions, lncRNA_CNV_profiles)
}
