% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_clin.R
\name{input_clin}
\alias{input_clin}
\title{Preparing input files for clinical data}
\usage{
input_clin(file_path = NULL)
}
\arguments{
\item{file_path}{Optional character string. If provided, a custom path to a tab-delimited
clinical file will be used. If not provided, the default system file
will be used from the package.}
}
\value{
A data frame of selected clinical data
}
\description{
Preparing input files for clinical data
}
\examples{
clin_df <- input_clin()  # Uses the default system file
clin_df <- input_clin("/custom/path/to/clinical_data.txt.gz")
}
