% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cnv.R
\name{input_cnv}
\alias{input_cnv}
\title{Preparing input files for CNV data}
\usage{
input_cnv(file_path = NULL)
}
\arguments{
\item{file_path}{Optional character string. If provided, a custom path to a tab-delimited
CNV file will be used. If not provided, the default system file
will be used from the package.}
}
\value{
A data frame of absolute CNV values
}
\description{
Preparing input files for CNV data
}
\examples{
cnv_df <- input_cnv()  # Uses the default system file
cnv_df <- input_cnv("/custom/path/to/CNV_data.txt.gz")
}
