% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_expr.R
\name{input_expr}
\alias{input_expr}
\title{Preparing input files for expression data}
\usage{
input_expr(file_path = NULL)
}
\arguments{
\item{file_path}{Optional character string. If provided, a custom path to a tab-delimited
expression file will be used. If not provided, the default system file
will be used from the package.}
}
\value{
A data frame of expression values
}
\description{
Preparing input files for expression data
}
\examples{
expr_df <- input_expr()  # Uses the default system file
expr_df <- input_expr("/custom/path/to/gene_expression.txt.gz")
}
