% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_lncRNA_positions.R
\name{input_lncRNA_positions}
\alias{input_lncRNA_positions}
\title{Reading the lncRNA positions in the human genome}
\usage{
input_lncRNA_positions(file_path = NULL)
}
\arguments{
\item{file_path}{Optional character string. If provided, a custom path to the
lncRNA positions file will be used. If not provided, the default
system file will be loaded.}
}
\value{
A data frame of lncRNA positions
}
\description{
Reading the lncRNA positions in the human genome
}
\examples{
lncRNA_positions <- input_lncRNA_positions()
lncRNA_positions <- input_lncRNA_positions("/custom/path/to/lncRNA_positions.txt")
}
