% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk_score_model_development.R
\name{Risk_score_model_development}
\alias{Risk_score_model_development}
\title{Extracting key lncRNAs and development of risk score model for them}
\usage{
Risk_score_model_development(
  result_surv = NULL,
  result_visualization = NULL,
  Correlation_lncRNAs_and_CNVs_results = NULL,
  merged_df_fetched_lncRNAs_expression = NULL,
  significant_lncRNA_names_after_violin_plot = NULL,
  final_matrix_corr_details = NULL
)
}
\arguments{
\item{result_surv}{list returned by extract_survival_related_significant_lncRNA}

\item{result_visualization}{list returned by visualization_of_lncRNA_and_CNVs_association}

\item{Correlation_lncRNAs_and_CNVs_results}{list returned by Correlation_lncRNAs_and_CNVs}

\item{merged_expression_clinical_data}{data frame of total lncRNAs expression values with clinical info}

\item{significant_lncRNA_names_from_kruskal}{data frame of lncRNAs corresponding CNV details}
}
\value{
result_riskscore_model
}
\description{
Extracting key lncRNAs and development of risk score model for them
}
\examples{
# Example 1: using the results from previous functions:
# Assuming result_surv and result_visualization are obtained from the previous functions
result_riskscore_model <- Risk_score_model_development(result_surv, result_visualization, Correlation_lncRNAs_and_CNVs_results)

# Example 2: providing the data frames directly:
# Assuming merged_df_fetched_lncRNAs_expression_df and significant_lncRNA_names_after_violin_plot_df are available
result_riskscore_model <- Risk_score_model_development(merged_df_fetched_lncRNAs_expression, significant_lncRNA_names_after_violin_plot, final_matrix_corr_details)
}
